Rename git into git-span-metadata
