+++ src/helpers/git-span-metadata.ts
